refactor bpod_settings into bpod_paths
