+++ bpod_rig/config/bpod_paths.py
